Исправляем замечания ревьювера.
